Replace AmplifyField and AmplifyFormField functionality with combined AmplifyFormField
