+++ packages/amplify-ui-components/src/components/amplify-examples/readme.md
@@ -9,8 +9,8 @@
 
 ### Depends on
 
-- [amplify-field](../amplify-field)
-- [amplify-field](../amplify-field)
+- [amplify-form-field](../amplify-form-field)
+- [amplify-form-field](../amplify-form-field)
 - [amplify-button](../amplify-button)
 - [amplify-link](../amplify-link)
 - [amplify-authenticator](../amplify-authenticator)
@@ -24,8 +24,8 @@
 ### Graph
 ```mermaid
 graph TD;
-  amplify-examples --> amplify-field
-  amplify-examples --> amplify-field
+  amplify-examples --> amplify-form-field
+  amplify-examples --> amplify-form-field
   amplify-examples --> amplify-button
   amplify-examples --> amplify-link
   amplify-examples --> amplify-authenticator
@@ -35,7 +35,8 @@ graph TD;
   amplify-examples --> amplify-authenticator
   amplify-examples --> rock-paper-scissor
   amplify-examples --> rock-paper-scissor
-  amplify-field --> amplify-input
+  amplify-form-field --> amplify-input
+  amplify-form-field --> amplify-hint
   amplify-authenticator --> amplify-sign-in
   amplify-authenticator --> context-consumer
   amplify-sign-in --> amplify-section
@@ -45,9 +46,6 @@ graph TD;
   amplify-sign-in --> amplify-button
   amplify-sign-in-username-field --> amplify-form-field
   amplify-sign-in-username-field --> context-consumer
-  amplify-form-field --> amplify-label
-  amplify-form-field --> amplify-input
-  amplify-form-field --> amplify-hint
   amplify-sign-in-password-field --> amplify-form-field
   amplify-sign-in-password-field --> context-consumer
   rock-paper-scissor --> amplify-button
