--- packages/amplify-ui-components/src/components/amplify-field/amplify-field.stories.tsx
@@ -1,13 +1,0 @@
-import { storiesOf } from '@storybook/html';
-import { text as textKnob } from '@storybook/addon-knobs';
-
-storiesOf('amplify-field', module).add('with knobs', () => {
-  const label = textKnob('Label', 'Label placeholder');
-  const description = textKnob('Description', 'Description placeholder');
-  const type = textKnob('Type', 'text');
-  return `<amplify-field field-id="id" label="${label}" description="${description}" type="${type}"/>`;
-});
-
-storiesOf('amplify-field', module).add('password example', () => {
-  return `<amplify-field field-id="id" label="Enter your password" description="Passwords should be kept safe!" type="password"/>`;
-});
