+++ packages/amplify-ui-components/src/components/amplify-form-field/amplify-form-field.spec.ts
@@ -1,11 +1,11 @@
 import { newSpecPage } from '@stencil/core/testing';
-import { AmplifyField } from './amplify-field.tsx'
+import { AmplifyFormField } from './amplify-form-field.tsx'
 
-describe('amplify-field', () => {
+describe('amplify-form-field', () => {
   it('renders no label or description if none are provided', async () => {
     const page = await newSpecPage({
-      components: [AmplifyField],
-      html: `<amplify-field></amplify-field>`,
+      components: [AmplifyFormField],
+      html: `<amplify-form-field></amplify-form-field>`,
     });
 
     expect(page.root).toMatchSnapshot();
@@ -13,8 +13,8 @@ describe('amplify-field', () => {
 
   it('renders a label and description, if they are provided, and no id', async () => {
     const page = await newSpecPage({
-      components: [AmplifyField],
-      html: `<amplify-field label='label' description='description'></amplify-field>`,
+      components: [AmplifyFormField],
+      html: `<amplify-form-field label='label' description='description'></amplify-form-field>`,
     });
 
     expect(page.root).toMatchSnapshot();
@@ -22,8 +22,8 @@ describe('amplify-field', () => {
 
   it('renders with an id, if it is provided', async () => {
     const page = await newSpecPage({
-      components: [AmplifyField],
-      html: `<amplify-field label='label' description='description' field-id='id'></amplify-field>`,
+      components: [AmplifyFormField],
+      html: `<amplify-form-field label='label' description='description' field-id='id'></amplify-form-field>`,
     });
 
     expect(page.root).toMatchSnapshot();
@@ -31,8 +31,8 @@ describe('amplify-field', () => {
 
   it('replaces the input component, if a new one is provided', async () => {
     const page = await newSpecPage({
-      components: [AmplifyField],
-      html: `<amplify-field label='label' description='description' field-id='id'><input slot="input" /></amplify-field>`,
+      components: [AmplifyFormField],
+      html: `<amplify-form-field label='label' description='description' field-id='id'><input slot="input" /></amplify-form-field>`,
     });
 
     expect(page.root).toMatchSnapshot();
