--- packages/amplify-ui-components/src/components/amplify-form-field/amplify-form-field.style.ts
@@ -1,5 +1,0 @@
-import { css } from 'emotion';
-
-export const formField = css`
-  margin-bottom: 22px;
-`;
