+++ packages/amplify-ui-components/src/components/amplify-label/readme.md
@@ -12,19 +12,6 @@
 | `htmlFor` | `html-for` |             | `string` | `undefined` |
 
 
-## Dependencies
-
-### Used by
-
- - [amplify-form-field](../amplify-form-field)
-
-### Graph
-```mermaid
-graph TD;
-  amplify-form-field --> amplify-label
-  style amplify-label fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
