+++ packages/amplify-ui-components/src/components/amplify-sign-in-password-field/readme.md
@@ -34,7 +34,6 @@
 graph TD;
   amplify-sign-in-password-field --> amplify-form-field
   amplify-sign-in-password-field --> context-consumer
-  amplify-form-field --> amplify-label
   amplify-form-field --> amplify-input
   amplify-form-field --> amplify-hint
   amplify-examples --> amplify-sign-in-password-field
