+++ packages/amplify-ui-components/src/components/amplify-sign-in/readme.md
@@ -37,7 +37,6 @@ graph TD;
   amplify-sign-in --> amplify-button
   amplify-sign-in-username-field --> amplify-form-field
   amplify-sign-in-username-field --> context-consumer
-  amplify-form-field --> amplify-label
   amplify-form-field --> amplify-input
   amplify-form-field --> amplify-hint
   amplify-sign-in-password-field --> amplify-form-field
